Added unittest for linear models + refactor to do it
