+++ technical_validation/experiments/match_mismatch_dilated_convolutional_model.py
@@ -8,7 +8,8 @@ os.environ['TF_XLA_FLAGS'] = "--tf_xla_auto_jit=2 --tf_xla_cpu_global_jit"
 import tensorflow as tf
 
 from technical_validation.models.dilated_convolutional_model import dilation_model
-from technical_validation.util.dataset_generator import MatchMismatchDataGenerator, default_batch_equalizer_fn, create_tf_dataset
+from technical_validation.util.dataset_generator_tf import default_batch_equalizer_fn, \
+    create_tf_dataset, MatchMismatchDataGenerator
 
 
 def evaluate_model(model, test_dict):
