+++ technical_validation/experiments/regression_vlaai.py
@@ -7,8 +7,8 @@ import os
 import tensorflow as tf
 
 from technical_validation.models.vlaai import vlaai, pearson_loss, pearson_metric
-from technical_validation.util.dataset_generator import RegressionDataGenerator, \
-    create_tf_dataset
+from technical_validation.util.dataset_generator import RegressionDataGenerator
+from technical_validation.util.dataset_generator_tf import create_tf_dataset
 
 
 def evaluate_model(model, test_dict):
